Introduced a business layer
The new business layer can have our application logic. Needed since neither the controller nor the repository should hold business logic.

Separated our services in a CQRS style, one interface for the commands/writes and another for the queries/reads.

For now most of our methods are just proxies in the business layer but when we have more logic we have the proper place to put it